Removed warnings

+++ AnotherOneBitesTheCrust/ContentLoading.cpp
@@ -441,44 +441,44 @@ bool ContentLoading::loadMap(char* filename, Map &map) {
 
 			// Handle rotations
 			if (rotation == "R") {
-				for (int i = 0; i < tile.entities.size(); i++) {
-					int x = tile.entities[i].position.x;
-					int z = tile.entities[i].position.z;
-					tile.entities[i].position.x = tileSize - z;
-					tile.entities[i].position.z = x;
+				for (unsigned int i = 0; i < tile.entities.size(); i++) {
+					int x = (int)tile.entities[i].position.x;
+					int z = (int)tile.entities[i].position.z;
+					tile.entities[i].position.x = (float)(tileSize - z);
+					tile.entities[i].position.z = (float)x;
 				}
-				for (int i = 0; i < tileNodes.size(); i++) {
+				for (unsigned int i = 0; i < tileNodes.size(); i++) {
 					glm::vec3 pos = tileNodes[i].position;
-					pos.x = 1.0 - pos.z;
+					pos.x = 1.0f - pos.z;
 					pos.z = pos.x;
 					tileNodes[i].position = pos;
 				}
 			}
 			if (rotation == "L") {
-				for (int i = 0; i < tile.entities.size(); i++) {
-					int x = tile.entities[i].position.x;
-					int z = tile.entities[i].position.z;
-					tile.entities[i].position.x = z;
-					tile.entities[i].position.z = tileSize - x;
+				for (unsigned int i = 0; i < tile.entities.size(); i++) {
+					int x = (int)tile.entities[i].position.x;
+					int z = (int)tile.entities[i].position.z;
+					tile.entities[i].position.x = (float)z;
+					tile.entities[i].position.z = (float)(tileSize - x);
 				}
-				for (int i = 0; i < tileNodes.size(); i++) {
+				for (unsigned int i = 0; i < tileNodes.size(); i++) {
 					glm::vec3 pos = tileNodes[i].position;
 					pos.x = pos.z;
-					pos.z = 1.0 - pos.x;
+					pos.z = 1.0f - pos.x;
 					tileNodes[i].position = pos;
 				}
 			}
 			if (rotation == "RR" || rotation == "LL") {
-				for (int i = 0; i < tile.entities.size(); i++) {
-					int x = tile.entities[i].position.x;
-					int z = tile.entities[i].position.z;
-					tile.entities[i].position.x = tileSize - x;
-					tile.entities[i].position.z = tileSize - z;
+				for (unsigned int i = 0; i < tile.entities.size(); i++) {
+					int x = (int)tile.entities[i].position.x;
+					int z = (int)tile.entities[i].position.z;
+					tile.entities[i].position.x = (float)(tileSize - x);
+					tile.entities[i].position.z = (float)(tileSize - z);
 				}
-				for (int i = 0; i < tileNodes.size(); i++) {
+				for (unsigned int i = 0; i < tileNodes.size(); i++) {
 					glm::vec3 pos = tileNodes[i].position;
-					pos.x = 1.0 - pos.x;
-					pos.z = 1.0 - pos.z;
+					pos.x = 1.0f - pos.x;
+					pos.z = 1.0f - pos.z;
 					tileNodes[i].position = pos;
 				}
 			}
@@ -498,10 +498,10 @@ bool ContentLoading::loadMap(char* filename, Map &map) {
 			}
 
 			// Local Connections
-			for(int k = 0; k < tileNodes.size(); k++)
+			for(unsigned int k = 0; k < tileNodes.size(); k++)
 			{
 				NodeTemplate current = tileNodes.at(k);
-				for(int l = 0; l < current.neighbours.size(); l++)
+				for(unsigned int l = 0; l < current.neighbours.size(); l++)
 				{
 					int index = current.neighbours[l];
 					tile.nodes.at(k)->addNeighbour(tile.nodes.at(index));
